+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New feature to init generated applet as git repository (through wizard or `--git yes`) when `git` command is present on machine
+
 ## [1.9.1] - 2025-03-20
 ### Fixed
 - Fixed `escheck` npm command by adding npx